Training update (#152)
* build: ⬆️ Pin latests release

* build: ⬆️ Update nightly version

* feat: ✨ Update wifi module

* feat: ✨ Update led module

* feat: ✨ Add examples

* feat: ✨ Initial update of temp_sensors module

* chore: 🚧 Comment temp_sensor module

* feat: ✨ Update hardware-check example

* build: 🔥 Remove unnecesary dependencies

* build: 🔥 Remove unnecesary dependencies

* docs: 📝 Add todos

* fix: ⚡️ Updarte clock_divider

* build: 🔧 Update nightly toolchain

* build: 🔧 Update custom runner

* build: 🔧 Update nightly toolchain

* feat: ✨ Update http client example

* chore: 🏗️ Update cargo.lock

* feat: ✨ Initial update of http client examples

* feat: ✨ Update hhtp client example

* feat: ✨ Update hhtp client example

* feat: ✨ Update hhtps client example

* feat: ✨ Update hhtp client exercise

* feat: ✨ Update hhtp client exercise

* feat: ⚡️ Clippy check

* style: 🎨 Update imports

* feat: 🎨 Format includes

* build: ⬆️ Update rust edition

* feat: 🎨 Format includes

* feat: ✨ Initial update of http server

* feat: ✨ Initial update of button interrupt

* feat: ✨ Update http server

* build: ⬆️ Update dependencies

* build: ⬆️ Update dependencies

* build: ⬆️ Update dependencies

* build: ⬆️ Update dependencies

* build: 🏗️ Update authors

* build: ⬆️ Update dependencies

* build: ⬆️ Upgrade dependencies

* fix: 🐛 Fix led pin

* feat: ✨ Update mqtt exercise

* feat: ✨ Update i2c-sensor-reading exercise

* style: 🎨 Improve format

* style: 🎨 Improve format

* style: 🎨 Improve format

* style: 🎨 Improve format

* fix: 🐛 Fix typo

* feat: ✨ Update button interrupt example

* docs: 🎨 Update format

* docs: 🎨 Update format and links

* docs: 🎨 Update format

* docs: ✨ Update book.toml

* docs: 📝 Update software chapter

* docs: 📝 Update repository chapter

* ci: ⚡️ Update ci conditions

* docs: 🐛 Remove installing target

* docs: 📝 Update Hello, board section

* docs: 🔥 Remove readme

* docs: 📝 Add intro chapter

* build: ➖ Remove esp32c3 crate

* style: ⚡️ Remove unnecesary mut

* build: ➖ Remove unneded crates

* docs: 📝 Update project organization

* docs: 📝 Update cargo-generate chapter

* docs: 📝 Update instructions

* style: 🎨 Format logs

* style: 🎨 Format logs

* docs: 📝 Update execise content

* docs: 📝 Update https client instructions

* docs: 🎨 Format code snipets md lang

* docs: 📝 Remove todo

* docs: 🎨 Remove $ from console logs where its not needed

* docs: 🎨 Format summary

* docs: 🐛 Fix links

* chore: 🔥 Remove unused md file

* feat: ✨ Update http server instructions

* docs: 📝 Update initial mqtt chapters

* docs: 📝 Add todo

* docs: 🎨 Update mqtt instructions

* docs: 🎨 Remove unused mark

* docs: ✨ Introduce hint mark

* docs: ✨ Update warning mark

* docs: 📝 Add rust on esp reference

* docs: 📝 Update mqtt part 2 instructions

* chore: 🎨 Format log prints

* docs: 🎨 LanguageTool suggestions

* docs: 🎨 Standarize titles

* docs: 📝 Add marks

* docs: 📝 Update format and LT suggestions

* docs: 🎨 Format ESP-IDF naming

* docs: 📝 Update LowLevel chapter

* docs: 📝 Update i2c chapter

* docs: 📝 Update part 1 instructions

* docs: 📝 Update part 2 instructions

* Remove local imc42670 crate

* build: ⬆️ Update crates versions

* docs: 📝 Update datasheet links

* docs: 📝 Update easy version instructions

* docs: 📝 Update docstrings

* docs: 📝 Update i2c-driver hard version

* style: 🎨 Rename i2c exercises in cargo.toml

* docs: 📝 LanguageTool suggestions

* docs: 📝 LanguageTool suggestions

* docs: 📝 Fix tyoi

* docs: 📝 Update interrup instructions

* docs: 📝 Update interrup instructions

* docs: 📝 Update reference chapter

* docs: 📝 LanguageTool warnings

* feat: ✨ Add rgb-led crate

* feat: ✨ Add wifi crate

* feat: 🔥 Remove bsc

* feat: ✨ Remove bsc crate

* feat: ✨ Update examples to use rgb_led and wifi crates

* fix: 🐛 Update crates path

* docs: 📝 Update advanced readme

* docs: 📝 Update docstring

* docs: 📝 Remove Windows note

* docs: ⚡️ Remove TODOs

* build: 🐛 Fix dependencies

* chore: 📝 Update authors

* docs: 🎨 Format comments

+++ advanced/i2c-driver/src/icm42670p_solution.rs
@@ -3,7 +3,7 @@
 use embedded_hal::blocking::i2c;
 
 /// ICM42670P device driver.
-/// Datasheet: https://3cfeqx1hf82y3xcoull08ihx-wpengine.netdna-ssl.com/wp-content/uploads/2021/07/DS-000451-ICM-42670-P-v1.0.pdf
+/// Datasheet: https://invensense.tdk.com/wp-content/uploads/2021/07/DS-000451-ICM-42670-P-v1.0.pdf
 #[derive(Debug)]
 pub struct ICM42670P<I2C> {
     // The concrete I²C device implementation.
@@ -23,7 +23,7 @@ pub enum DeviceAddr {
     AD1 = 0b110_1001,
 }
 
-impl<I2C, E>ICM42670P<I2C>
+impl<I2C, E> ICM42670P<I2C>
 where
     I2C: i2c::WriteRead<Error = E> + i2c::Write<Error = E>,
 {
@@ -43,7 +43,7 @@ where
     // This method is not public as it is only needed inside this file.
     #[allow(unused)]
     fn write_register(&mut self, register: Register, value: u8) -> Result<(), E> {
-        let byte = value as u8;
+        let byte = value;
         self.i2c
             .write(self.address as u8, &[register.address(), byte])
     }
